fix(audit): /code-audit PR 554 Pass 1 corrections
- ddccontroller: remove dead refresh() slot (no caller); migrate off the
  deprecated ddca_set_non_table_vcp_value to the _2 variant; make enumerate()
  self-enforcing against a re-enumeration leak (free prior list + clear refs);
  document why init checks status < 0 vs != 0.
- brightnesshost: require the colon-delimited form for kbd_backlight so a bare
  name is not matched; note ddcDevices is GUI-thread-only; drop an em-dash.
- brightnessdevice: correct the readMax / header docs that cited an unused
  actual_brightness read; log when a brightness watch is lost on re-add.
- brightnessdevicemodel: cast m_rows.size() to int in the countChanged guard.
- CLI: correct the stale header comment (DDC enumeration is async + slow).
- shell main: add the Brightness stanza to the registration doc block.
- tests: assert the zero-max write is inert; add applyExternalValue no-change /
  clamp coverage; add Keyboard -> "leds" subsystem write coverage; add model
  deviceAdded / deviceRemoved transaction and host-destroyed reset tests;
  strengthen the idempotency test to assert no duplicate-registration warning;
  drop the unused actual_brightness from the fake sysfs helper.
- docs: add id to the M3 role list, get/set <id> in M5, drop actual_brightness
  from M2 / M6, note the leds subsystem coverage.

+++ libs/phosphor-service-brightness/src/brightnessdevicemodel.cpp
@@ -56,7 +56,7 @@ void BrightnessDeviceModel::setHost(BrightnessHost* host)
     }
     endResetModel();
     Q_EMIT hostChanged();
-    if (previousCount != m_rows.size())
+    if (previousCount != static_cast<int>(m_rows.size()))
         Q_EMIT countChanged();
 }
